Move min/max into scope of CMIP6 chart
